Update README.md
A travel company wants to predict whether its customers will churn based on indicators like age, frequent flyer information, annual income, and services used. 
This project is focused on predicting customer churn for a tour and travels company. Customer churn refers to when a customer stops using a company's services. In this case, we want to predict when a customer might stop booking tours with our company.